minigames/requirements.txt
adrf==0.1.6
aiofiles==23.2.1
aiogram==3.8.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
async-property==0.2.2
attrs==23.2.0
certifi==2024.6.2
channels==4.1.0
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
frozenlist==1.4.1
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
psycopg2-binary==2.9.9
pydantic==2.7.4
pydantic_core==2.18.4
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.12.2
yarl==1.9.4
daphne
pillow
